Handle larger fonts

+++ src/epd_driver/include/epd_internals.h
@@ -113,12 +113,12 @@ extern const EpdWaveform epdiy_ED133UT2;
 
 /// Font data stored PER GLYPH
 typedef struct {
-  uint8_t width;            ///< Bitmap dimensions in pixels
-  uint8_t height;           ///< Bitmap dimensions in pixels
-  uint8_t advance_x;        ///< Distance to advance cursor (x axis)
+  uint16_t width;            ///< Bitmap dimensions in pixels
+  uint16_t height;           ///< Bitmap dimensions in pixels
+  uint16_t advance_x;        ///< Distance to advance cursor (x axis)
   int16_t left;             ///< X dist from cursor pos to UL corner
   int16_t top;              ///< Y dist from cursor pos to UL corner
-  uint16_t compressed_size; ///< Size of the zlib-compressed font data.
+  uint32_t compressed_size; ///< Size of the zlib-compressed font data.
   uint32_t data_offset;     ///< Pointer into EpdFont->bitmap
 } EpdGlyph;
